requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scikit-learn>=1.4
scikit-image>=0.22
scipy>=1.11
pyyaml>=6.0
rich>=13.0
matplotlib>=3.8
seaborn>=0.13
plotly>=5.20

# Deep learning
torch>=2.1
torchvision>=0.16
timm>=0.9
albumentations>=1.4
opencv-python-headless>=4.9

# NLP / VLM integration
transformers>=4.40
accelerate>=0.30
safetensors>=0.4
openai>=1.40
huggingface-hub>=0.23
fastapi>=0.111
uvicorn>=0.29

# Experiment management
hydra-core>=1.3
omegaconf>=2.3
torchmetrics>=1.4

# Reporting
jupyterlab>=4.0
polars>=0.20
pydantic>=2.6

# Optional fairness tooling
fairlearn>=0.10
dusql>=0.0.7

